+++ node/resolvers/pageResolvers.ts
@@ -11,7 +11,7 @@ export const pageResolvers = {
     try {
       return await wordpressProxy.getUser(author, customDomain)
     } catch (e) {
-      return console.error(`${e.name}: ${e.message}`)
+      console.error(`${e.name}: ${e.message}`)
     }
   },
   featured_media: async (
@@ -29,10 +29,9 @@ export const pageResolvers = {
       try {
         return await wordpressProxy.getMediaSingle(featured_media, customDomain)
       } catch (e) {
-        return console.error(`${e.name}: ${e.message}`)
+        console.error(`${e.name}: ${e.message}`)
       }
-    } else {
-      return null
     }
+    return null
   },
 }