+++ react/components/WordpressCategory.tsx
@@ -43,7 +43,6 @@ const WordpressCategory: StorefrontFunctionComponent<CategoryProps> = ({
     parsedCustomDomains = customDomains ? JSON.parse(customDomains) : null
   } catch (e) {
     console.error(`${e.name}: ${e.message}`)
-    return null
   }
 
   const customDomain =