+++ react/components/WordpressPost.tsx
@@ -312,7 +312,6 @@ const WordpressPost: StorefrontFunctionComponent<PostProps> = ({
     parsedCustomDomains = customDomains ? JSON.parse(customDomains) : null
   } catch (e) {
     console.error(`${e.name}: ${e.message}`)
-    return null
   }
 
   const customDomain =